MODCON-25 fixed comments
