+++ src/main/java/org/folio/consortia/service/impl/TenantServiceImpl.java
@@ -3,23 +3,16 @@ package org.folio.consortia.service.impl;
 import static org.folio.consortia.utils.HelperUtils.checkIdenticalOrThrow;
 import static org.folio.consortia.utils.TenantContextUtils.prepareContextForTenant;
 
-import java.util.Date;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.IntStream;
 
-import org.apache.commons.lang3.StringUtils;
 import org.folio.consortia.client.ConsortiaConfigurationClient;
-import org.folio.consortia.client.UsersClient;
-import org.folio.consortia.config.kafka.KafkaService;
 import org.folio.consortia.domain.dto.ConsortiaConfiguration;
 import org.folio.consortia.domain.dto.PermissionUser;
 import org.folio.consortia.domain.dto.Tenant;
 import org.folio.consortia.domain.dto.TenantCollection;
 import org.folio.consortia.domain.dto.User;
-import org.folio.consortia.domain.dto.UserEvent;
 import org.folio.consortia.domain.entity.TenantEntity;
 import org.folio.consortia.domain.entity.UserTenantEntity;
 import org.folio.consortia.exception.ResourceAlreadyExistException;
@@ -29,8 +22,8 @@ import org.folio.consortia.repository.UserTenantRepository;
 import org.folio.consortia.service.ConsortiumService;
 import org.folio.consortia.service.PermissionUserService;
 import org.folio.consortia.service.TenantService;
+import org.folio.consortia.service.UserAffiliationAsyncService;
 import org.folio.consortia.service.UserService;
-import org.folio.consortia.service.UserTenantService;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.scope.FolioExecutionContextSetter;
@@ -62,9 +55,7 @@ public class TenantServiceImpl implements TenantService {
   private final PermissionUserService permissionUserService;
   private final UserService userService;
   private final FolioModuleMetadata folioModuleMetadata;
-  private final UsersClient usersClient;
-  private final UserTenantService userTenantService;
-  private final KafkaService kafkaService;
+  private final UserAffiliationAsyncService createPrimaryUserAffiliationsAsync;
 
   @Override
   public TenantCollection get(UUID consortiumId, Integer offset, Integer limit) {
@@ -114,50 +105,16 @@ public class TenantServiceImpl implements TenantService {
 
     try (var context = new FolioExecutionContextSetter(prepareContextForTenant(tenantDto.getId(), folioModuleMetadata, currentTenantContext))) {
       createShadowAdminUserWithPermissions(shadowAdminUser);
-      createPrimaryUserAffiliationsAsync(consortiumId, savedTenantEntity, tenantDto, currentTenantContext.getUserId());
+      createPrimaryUserAffiliationsAsync.createPrimaryUserAffiliationsAsync(consortiumId, savedTenantEntity, tenantDto, currentTenantContext.getUserId());
       configurationClient.saveConfiguration(createConsortiaConfigurationBody(centralTenantId));
     }
     log.info("save:: saved consortia configuration with centralTenantId={} by tenantId={} context", centralTenantId, tenantDto.getId());
     return savedTenant;
   }
 
-  public CompletableFuture<Void> createPrimaryUserAffiliationsAsync(UUID consortiumId, TenantEntity consortiaTenant,
-      Tenant tenantDto, UUID contextUserId) {
-    return CompletableFuture.runAsync(() -> {
-      log.info("Start creating user primary affiliation for tenant {}", tenantDto.getId());
-      var users = usersClient.getUserCollection(StringUtils.EMPTY, 0, Integer.MAX_VALUE);
-      log.info("{} tenant users found", users.getUsers().size());
-      IntStream.range(0, users.getUsers().size())
-        .forEach(idx -> {
-          var user = users.getUsers().get(idx);
-          log.info("Processing users: {} of {}", idx + 1, users.getUsers().size());
-          var consortiaUserTenant = userTenantRepository.findByUserIdAndTenantId(UUID.fromString(user.getId()), tenantDto.getId())
-            .orElse(null);
-          if (consortiaUserTenant != null && consortiaUserTenant.getIsPrimary()) {
-            log.info("Primary affiliation already exists for tenant/user: {}/{}", tenantDto.getId(), user.getUsername());
-          } else {
-            userTenantService.createPrimaryUserTenantAffiliation(consortiumId, consortiaTenant, user.getId(), user.getUsername());
-            sendCreatePrimaryAffiliationEvent(consortiaTenant, tenantDto, contextUserId, user);
-          }
-        });
-    })
-      .thenAccept(v -> log.info("Successfully created primary affiliations for tenant {}", tenantDto.getId()))
-      .exceptionally(t -> {
-        log.error("Failed to create primary affiliations for new tenant", t);
-        return null;
-      });
-  }
 
-  private void sendCreatePrimaryAffiliationEvent(TenantEntity consortiaTenant, Tenant tenantDto, UUID contextUserId, User user) {
-    var ue = new UserEvent().tenantId(tenantDto.getId())
-      .userDto(user)
-      .action(UserEvent.ActionEnum.CREATE)
-      .actionDate(new Date())
-      .eventDate(new Date())
-      .performedBy(contextUserId);
 
-    kafkaService.send(KafkaService.Topic.CONSORTIUM_PRIMARY_AFFILIATION_CREATED, consortiaTenant.getConsortiumId().toString(), ue.toString());
-  }
+
 
   @Override
   public Tenant update(UUID consortiumId, String tenantId, Tenant tenantDto, Boolean forceCreatePrimaryAff) {
@@ -169,7 +126,7 @@ public class TenantServiceImpl implements TenantService {
 
     if (forceCreatePrimaryAff.booleanValue()) {
       try (var context = new FolioExecutionContextSetter(prepareContextForTenant(tenantDto.getId(), folioModuleMetadata, currentTenantContext))) {
-        createPrimaryUserAffiliationsAsync(consortiumId, tenantEntity, tenantDto, currentTenantContext.getUserId());
+        createPrimaryUserAffiliationsAsync.createPrimaryUserAffiliationsAsync(consortiumId, tenantEntity, tenantDto, currentTenantContext.getUserId());
       }
     }
     return savedTenant;
